requirements.txt
flask_sqlalchemy
flask_marshmallow
werkzeug